#125: Extracted helper functions into utils
