+++ test/utils/result_utils.py
@@ -1,0 +1,33 @@
+import json
+from typing import (
+    Any,
+    cast,
+)
+
+from exasol.ai.mcp.server.server_settings import ExaDbResult
+
+
+def result_sort_func(d: Any) -> str:
+    if isinstance(d, dict):
+        return ",".join(str(d[key]) for key in sorted(d.keys()))
+    return str(d)
+
+
+def get_result_json(result) -> dict[str, Any]:
+    return cast(dict[str, Any], json.loads(result.content[0].text))
+
+
+def get_sort_result_json(result) -> dict[str, Any]:
+    result_json = get_result_json(result)
+    return {
+        key: sorted(val, key=result_sort_func) if isinstance(val, list) else val
+        for key, val in result_json.items()
+    }
+
+
+def get_list_result_json(result) -> ExaDbResult:
+    result_json = get_result_json(result)
+    unsorted = ExaDbResult(**result_json)
+    if isinstance(unsorted.result, list):
+        return ExaDbResult(sorted(unsorted.result, key=result_sort_func))
+    return unsorted
